[Dx12] Added support for UploadCommandList, Vertex Buffers and Index Buffers

+++ Arc/src/Platform/OpenGL/OpenGLBuffer.cpp
@@ -10,22 +10,22 @@ namespace ArcEngine
 	// VertexBuffer //////////////////////////////////////////////////////
 	//////////////////////////////////////////////////////////////////////
 
-	OpenGLVertexBuffer::OpenGLVertexBuffer(size_t size)
+	OpenGLVertexBuffer::OpenGLVertexBuffer(uint32_t size)
 	{
 		ARC_PROFILE_SCOPE()
 
 		glCreateBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
-		glBufferData(GL_ARRAY_BUFFER, static_cast<int>(size), nullptr, GL_DYNAMIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
 	}
 
-	OpenGLVertexBuffer::OpenGLVertexBuffer(const float* vertices, size_t size)
+	OpenGLVertexBuffer::OpenGLVertexBuffer(const float* vertices, uint32_t size)
 	{
 		ARC_PROFILE_SCOPE()
 		
 		glCreateBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
-		glBufferData(GL_ARRAY_BUFFER, static_cast<int>(size), vertices, GL_STATIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, size, vertices, GL_STATIC_DRAW);
 	}
 
 	OpenGLVertexBuffer::~OpenGLVertexBuffer()
